Refactor backend authentication file structure for better organization
