+++ src/backend/index.js
@@ -20,6 +20,10 @@ import { Server } from 'socket.io';
 import db from './config/db.js';
 import supervisorRoutes from './routes/supervisorRoutes.js';
 import inviteRoutes from './routes/inviteRoutes.js';
+import authRoutes from './routes/authRoutes.js';
+import authLoginRoutes from './routes/authLoginRoutes.js'; 
+import authPasswordRoutes from './routes/authPasswordRoutes.js';
+import authResetRoutes from './routes/authResetRoutes.js';
 
 const app = express();
 app.use(bodyParser.json());
@@ -103,58 +107,9 @@ const saltRounds = 10;
 
 // -------Register endpoint------- //
 
-app.post("/register", async (req, res) => {
-  const { FullName, Email, Password, Role, Phone } = req.body;
+app.use(authRoutes); 
 
-  try {
-    // Hash the password
-    const hashedPassword = await bcrypt.hash(Password, 10);
-
-    // Insert the user into appuser
-    const query = 'INSERT INTO appuser (FullName, Email, Password, Role, Phone) VALUES (?, ?, ?, ?, ?)';
-
-    db.query(query, [FullName, Email, hashedPassword, Role, Phone], async (err, result) => {
-      if (err) {
-        console.error('Error inserting user:', err);
-        return res.status(500).send('Error registering user');
-      }
-
-      const newUserID = result.insertId;
-
-      // Optional: Link the user to an existing client with matching email
-      const updateClientQuery = `
-        UPDATE client 
-        SET UserID = ? 
-        WHERE ContactPersonEmail = ? AND UserID IS NULL
-      `;
-
-      db.query(updateClientQuery, [newUserID, Email], async (updateErr, updateResult) => {
-        if (updateErr) {
-          console.error('Error updating client with UserID:', updateErr);
-          // Not fatal — continue registration
-        } else {
-          console.log(`Client records updated with UserID: ${newUserID}`);
-        }
-
-        // Send notification to admins
-        try {
-          await sendNotificationsByRoles(
-            ['Admin'],
-            `New ${Role} registered: ${FullName} (${Email})`,
-            'NEW_USER_REGISTRATION'
-          );
-        } catch (notificationError) {
-          console.error('Error sending registration notifications:', notificationError);
-        }
-
-        return res.status(200).send('User registered successfully');
-      });
-    });
-  } catch (error) {
-    console.error('Registration error:', error);
-    return res.status(500).send('Internal server error');
-  }
-});
+/*------------------------------------------*/
 
 // API endpoint to fetch tickets
 app.get('/api/tickets', (req, res) => {
@@ -177,48 +132,10 @@ app.get('/api/tickets', (req, res) => {
 });
 
 // Login endpoint
-app.post('/login', (req, res) => {
-  const { Email, Password } = req.body;
-  const query = 'SELECT UserID, FullName, Email, Phone, Role, ProfileImagePath, Password AS HashedPassword FROM appuser WHERE Email = ?';
 
-  db.query(query, [Email], (err, results) => {
-    if (err) {
-      console.error('Error during login:', err);
-      return res.status(500).json({ message: 'Error during login' });
-    }
+app.use(authLoginRoutes);
 
-    if (results.length === 0) {
-      return res.status(401).json({ message: 'Invalid credentials' });
-    }
-
-    const user = results[0];
-
-    bcrypt.compare(Password, user.HashedPassword, (compareErr, isMatch) => {
-      if (compareErr) {
-        console.error('Error comparing passwords:', compareErr);
-        return res.status(500).json({ message: 'Error during login' });
-      }
-
-      if (isMatch) {
-        // Passwords match! Login successful.
-        res.status(200).json({
-          message: 'Login successful',
-          user: {
-            UserID: user.UserID,
-            FullName: user.FullName,
-            Email: user.Email,
-            Phone: user.Phone,
-            Role: user.Role,
-            ProfileImagePath: user.ProfileImagePath || null
-          }
-        });
-      } else {
-        // Passwords do not match
-        res.status(401).json({ message: 'Invalid credentials' });
-      }
-    });
-  });
-});
+/*---------------------------------------------*/
 
 // Get admin  profile endpoint 
 app.get('/api/user/profile/:id', (req, res) => {
@@ -354,164 +271,14 @@ app.delete('/api/user/profile/image/:id', async (req, res) => {
 
 /*---------------------- Invite User via Email ----------------------*/
 
-
-
 // --- ADDED: FORGOT PASSWORD ENDPOINT ---
-app.post('/forgot-password', (req, res) => {
-  const { email } = req.body;
 
-  if (!email || !isValidEmail(email)) {
-    return res.status(400).json({ message: 'A valid email address is required.' });
-  }
+app.use('/', authPasswordRoutes);
 
-  // 1. Find the user by email
-  db.query('SELECT UserID, FullName FROM appuser WHERE Email = ?', [email], (err, users) => {
-    if (err) {
-      console.error('Error finding user for password reset:', err);
-      return res.status(500).json({ message: 'Server error. Please try again later.' });
-    }
-
-    if (users.length === 0) {
-      // For security, always send a generic success message even if email not found
-      return res.status(200).json({ message: 'If an account with that email exists, a password reset link has been sent.' });
-    }
-
-    const user = users[0];
-    const token = uuidv4(); // Generate a unique token
-    const expiresAt = moment().add(1, 'hour').format('YYYY-MM-DD HH:mm:ss'); // Token expires in 1 hour
-
-    // 2. Invalidate any existing tokens for this user
-    db.query('DELETE FROM password_reset_tokens WHERE user_id = ?', [user.UserID], (deleteErr) => {
-      if (deleteErr) {
-        console.error('Error deleting old tokens:', deleteErr);
-        return res.status(500).json({ message: 'Server error during token invalidation.' });
-      }
-
-      // 3. Store the new token in the database
-      const insertTokenQuery = 'INSERT INTO password_reset_tokens (user_id, token, expires_at) VALUES (?, ?, ?)';
-      db.query(insertTokenQuery, [user.UserID, token, expiresAt], (insertErr) => {
-        if (insertErr) {
-          console.error('Error storing new token:', insertErr);
-          return res.status(500).json({ message: 'Server error during token storage.' });
-        }
-
-        // 4. Send the password reset email
-        const resetLink = `${process.env.APP_URL}/reset-password?token=${token}&email=${encodeURIComponent(email)}`;
-
-        const mailOptions = {
-          from: `"Asipiya Soft Solution (PVT) LTD" <${process.env.EMAIL_USER}>`,
-          to: email,
-          subject: 'Password Reset Request',
-          html: `
-                        <div style="font-family: Arial, sans-serif; color: #333; max-width: 600px; margin: auto; background-color: #f0f8ff; border: 1px solid #ddd; border-radius: 10px; padding: 20px;">
-                            <div style="display: flex; align-items: center; gap: 10px; margin-bottom: 20px;">
-                                <img src="https://miro.medium.com/v2/resize:fit:2400/1*-okfOPV73mecuWxCZz6uJA.jpeg" alt="Asipiya Logo" style="width: 50px; height: auto;">
-                                <h2 style="margin: 0; color: #005baa;">Asipiya Soft Solution (PVT) LTD</h2>
-                            </div>
-
-                            <h3 style="text-align: center; color: #005baa;">Password Reset Request</h3>
-                            <p>Hello ${user.FullName || 'User'},</p>
-                            <p>We received a request to reset the password for your account. If you did not make this request, please ignore this email.</p>
-                            <p>To reset your password, please click the button below:</p>
-
-                            <div style="text-align: center; margin: 30px 0;">
-                                <a href="${resetLink}"
-                                   style="background: #005baa; color: #fff; text-decoration: none; padding: 14px 24px; border-radius: 6px; font-size: 16px;">
-                                    Reset Your Password
-                                </a>
-                            </div>
-
-                            <p style="font-size: 12px; color: #888;">
-                                This link will expire in 1 hour. If you do not reset your password within this timeframe, you will need to request a new link.
-                            </p>
-                            <p>If you have any questions, feel free to contact our support team.</p>
-
-                            <hr style="margin-top: 30px;">
-                            <p style="font-size: 12px; text-align: center; color: #aaa;">&copy; ${new Date().getFullYear()} Asipiya Soft Solution (PVT) LTD. All rights reserved.</p>
-                        </div>
-                    `
-        };
-
-        transporter.sendMail(mailOptions, (mailErr) => {
-          if (mailErr) {
-            console.error('Error sending password reset email:', mailErr);
-            return res.status(500).json({ message: 'Failed to send email' });
-          }
-          res.status(200).json({ message: 'If an account with that email exists, a password reset link has been sent.' });
-        });
-      });
-    });
-  });
-});
+/*----------------------------------------------*/
 
 // --- ADDED: RESET PASSWORD ENDPOINT ---
-app.post('/reset-password', (req, res) => {
-  const { email, token, newPassword } = req.body;
-
-  if (!email || !token || !newPassword) {
-    return res.status(400).json({ message: 'Email, token, and new password are required.' });
-  }
-
-  // 1. Find and validate the token
-  db.query('SELECT user_id, expires_at FROM password_reset_tokens WHERE token = ?', [token], (err, tokens) => {
-    if (err) {
-      console.error('Error finding token:', err);
-      return res.status(500).json({ message: 'Server error during token validation.' });
-    }
-
-    if (tokens.length === 0) {
-      return res.status(400).json({ message: 'Invalid or expired token. Please request a new password reset link.' });
-    }
-
-    const resetToken = tokens[0];
-    const currentDateTime = moment();
-    const expiryDateTime = moment(resetToken.expires_at);
-
-    if (currentDateTime.isAfter(expiryDateTime)) {
-      // Token has expired, delete it from the database
-      db.query('DELETE FROM password_reset_tokens WHERE token = ?', [token], (deleteErr) => {
-        if (deleteErr) console.error('Error deleting expired token:', deleteErr);
-        return res.status(400).json({ message: 'Token has expired. Please request a new password reset link.' });
-      });
-      return; // Exit here after sending response
-    }
-
-    // 2. Verify the user associated with the token matches the provided email
-    db.query('SELECT UserID FROM appuser WHERE UserID = ? AND Email = ?', [resetToken.user_id, email], (err, users) => {
-      if (err) {
-        console.error('Error verifying user for reset:', err);
-        return res.status(500).json({ message: 'Server error during user verification.' });
-      }
-
-      if (users.length === 0) {
-        return res.status(400).json({ message: 'Invalid user for this token.' });
-      }
-
-      // 3. Hash the new password
-      bcrypt.hash(newPassword, saltRounds, (hashErr, hashedPassword) => {
-        if (hashErr) {
-          console.error('Error hashing new password:', hashErr);
-          return res.status(500).json({ message: 'Error processing new password.' });
-        }
-
-        // 4. Update the user's password
-        db.query('UPDATE appuser SET Password = ? WHERE UserID = ?', [hashedPassword, resetToken.user_id], (updateErr) => {
-          if (updateErr) {
-            console.error('Error updating password:', updateErr);
-            return res.status(500).json({ message: 'Server error during password update.' });
-          }
-
-          // 5. Invalidate (delete) the used token from the database
-          db.query('DELETE FROM password_reset_tokens WHERE token = ?', [token], (deleteUsedTokenErr) => {
-            if (deleteUsedTokenErr) console.error('Error deleting used token:', deleteUsedTokenErr);
-            res.status(200).json({ message: 'Password has been reset successfully.' });
-          });
-        });
-      });
-    });
-  });
-});
-
+app.use('/', authResetRoutes);
 
 /*---------------------------------------------------------------------------------------*/
 // Socket.io connection
