+++ src/backend/routes/authLoginRoutes.js
@@ -1,0 +1,7 @@
+import express from 'express';
+import { login } from '../controllers/authLoginController.js';
+
+const router = express.Router();
+router.post('/login', login);
+
+export default router;
