+++ src/backend/routes/authRoutes.js
@@ -1,0 +1,8 @@
+// routes/authRoutes.js
+import express from 'express';
+import { register } from '../controllers/authController.js';
+
+const router = express.Router();
+router.post('/register', register);
+
+export default router;
